Refactor: 마이페이지 메인 리팩토링
