--- src/constants/myPage.ts
@@ -1,10 +1,0 @@
-const customerService = [
-  { title: '사용 가이드', url: '/mypage/guide' },
-  {
-    title: '자주 묻는 질문',
-    url: '/mypage/question',
-  },
-  { title: '식물 추가 요청', url: 'https://forms.gle/g4AjkNKqVDP48Xnc7' },
-];
-
-export { customerService };
